+++ packages/apollos-data-connector-rock/src/features/__tests__/index.js
@@ -43,9 +43,17 @@ describe('features', () => {
         },
       ],
     }));
-    first = jest.fn(() => ({ get: () => itemMock }));
+    first = jest.fn(() => Promise.resolve(itemMock));
     const byContentChannelId = () => ({
+      get: () => Promise.resolve(itemMock),
+      top: () => ({ get: () => Promise.resolve(itemMock) }),
+      first,
+    });
+    const getCursorByParentContentItemId = () => ({
       get: () => itemMock,
+      first,
+    });
+    const getSermonFeed = () => ({
       first,
     });
     context = {
@@ -53,6 +61,8 @@ describe('features', () => {
         ContentItem: {
           byPersonaFeed,
           byContentChannelId,
+          getCursorByParentContentItemId,
+          getSermonFeed,
           getCoverImage: () => null,
           resolveType: () => 'UniversalContentItem',
         },
@@ -103,6 +113,22 @@ describe('features', () => {
             arguments: { contentChannelId: 13 },
           },
         ],
+        title: 'Test Action List',
+        subtitle: "It's great!",
+      });
+
+      expect(result).toMatchSnapshot();
+      expect(first.mock.calls).toMatchSnapshot();
+    });
+
+    it('should create an ActionListFeature from a SERMON_CHILDREN algorithm', async () => {
+      const features = new Features();
+      features.initialize({
+        context,
+      });
+
+      const result = await features.createActionListFeature({
+        algorithms: ['SERMON_CHILDREN'],
         title: 'Test Action List',
         subtitle: "It's great!",
       });
@@ -201,10 +227,24 @@ describe('features', () => {
 
       expect(result).toMatchSnapshot();
     });
+
+    it('createScriptureFeature should create a Scriptre feature', async () => {
+      const features = new Features();
+      features.initialize({
+        context,
+      });
+
+      const result = await features.createScriptureFeature({
+        reference: 'John 3:16',
+        id: 123,
+      });
+
+      expect(result).toMatchSnapshot();
+    });
   });
 
   describe('resolver', () => {
-    it('must return a personaFeed and a contentChannelFeed for the userFeedFeatures', async () => {
+    it('must return a personaFeed, a sermonChildrenFeed, and a contentChannelFeed for the userFeedFeatures', async () => {
       const features = new Features();
       features.initialize({
         context,