+++ packages/apollos-rock-apollo-data-source/src/index.js
@@ -8,7 +8,7 @@ import { createCursor, parseCursor } from './cursor';
 
 import RequestBuilder from './request-builder';
 
-export { RockLoggingExtension } from './utils';
+export { RockLoggingExtension, parseKeyValueAttribute } from './utils';
 
 const { ROCK } = ApollosConfig;
 
@@ -40,7 +40,9 @@ export default class RockApolloDataSource extends RESTDataSource {
     }
     this.calls[request.path] = this.calls[request.path] + 1;
 
-    request.headers.set('Authorization-Token', this.rockToken);
+    if (!request.headers.has('Authorization-Token')) {
+      request.headers.set('Authorization-Token', this.rockToken);
+    }
     request.headers.set('user-agent', 'Apollos');
     request.headers.set('Content-Type', 'application/json');
   }