+++ packages/apollos-ui-kit/src/theme/defaultTheme.js
@@ -1,3 +1,5 @@
+import Color from 'color';
+
 import fontStack from './fontStack';
 import * as types from './types';
 
@@ -118,7 +120,7 @@ export const shadows = ({ colors: themeColors }) => ({
   },
 });
 
-export const buttons = ({ colors: themeColors }) => ({
+export const buttons = ({ colors: themeColors, alpha: themeAlpha }) => ({
   default: {
     fill: themeColors.action.default,
     accent: themeColors.text.primary,
@@ -139,10 +141,114 @@ export const buttons = ({ colors: themeColors }) => ({
     fill: themeColors.text.primary,
     accent: themeColors.text.primary,
   },
+  overlay: {
+    fill: Color(themeColors.white).alpha(themeAlpha.low),
+    accent: themeColors.white,
+  },
   alert: {
     fill: themeColors.alert,
     accent: themeColors.white,
   },
+});
+
+export const overlays = ({ alpha: themeAlpha, colors: themeColors }) => ({
+  // these are a function because we can't assume colors of an overlay based on a theme (varies on usage + context)
+  default: ({ overlayColor }) => ({
+    colors: [
+      `${Color(overlayColor)
+        .fade(1)
+        .string()}`,
+      overlayColor,
+    ],
+    start: { x: 0, y: 0 },
+    end: { x: 0, y: 1 },
+    locations: [0.5, 1],
+  }),
+  high: ({ overlayColor }) => ({
+    colors: [
+      `${Color(overlayColor)
+        .alpha(themeAlpha.high)
+        .string()}`,
+      `${Color(overlayColor)
+        .alpha(themeAlpha.high)
+        .string()}`,
+    ],
+    start: { x: 0, y: 0 },
+    end: { x: 0, y: 1 },
+    locations: [0, 1],
+  }),
+  medium: ({ overlayColor }) => ({
+    colors: [
+      `${Color(overlayColor)
+        .alpha(themeAlpha.medium)
+        .string()}`,
+      `${Color(overlayColor)
+        .alpha(themeAlpha.medium)
+        .string()}`,
+    ],
+    start: { x: 0, y: 0 },
+    end: { x: 0, y: 1 },
+    locations: [0, 1],
+  }),
+  low: ({ overlayColor }) => ({
+    colors: [
+      `${Color(overlayColor)
+        .alpha(themeAlpha.low)
+        .string()}`,
+      `${Color(overlayColor)
+        .alpha(themeAlpha.low)
+        .string()}`,
+    ],
+    start: { x: 0, y: 0 },
+    end: { x: 0, y: 1 },
+    locations: [0, 1],
+  }),
+  'gradient-bottom': ({ overlayColor }) => ({
+    colors: [
+      `${Color(overlayColor)
+        .alpha(themeAlpha.low)
+        .string()}`,
+      `${Color(overlayColor)
+        .alpha(themeAlpha.medium)
+        .string()}`,
+    ],
+    start: { x: 0, y: 0 },
+    end: { x: 0, y: 1 },
+    locations: [0, 1],
+  }),
+  'gradient-top': ({ overlayColor }) => ({
+    colors: [
+      `${Color(overlayColor)
+        .alpha(themeAlpha.medium)
+        .string()}`,
+      themeColors.transparent,
+    ],
+    start: { x: 0, y: 0 },
+    end: { x: 0, y: 1 },
+    locations: [0, 1],
+  }),
+  featured: ({ overlayColor }) => ({
+    colors: [
+      `${Color(overlayColor)
+        .alpha(themeAlpha.low)
+        .string()}`,
+      overlayColor,
+    ],
+    start: { x: 0, y: 0 },
+    end: { x: 0, y: 1 },
+    locations: [0, 0.8],
+  }),
+  // Overriding this property changes all BackgroundViews
+  'background-gradient': ({ colors: customColors }) => ({
+    colors: customColors || [
+      themeColors.background.paper,
+      themeColors.background.screen,
+    ],
+    // default props from `react-native-linear-gradient`
+    start: { x: 0.5, y: 0.0 },
+    end: { x: 0.5, y: 1.0 },
+    locations: null,
+  }),
 });
 
 /*